+++ docs/conf.py
@@ -48,8 +48,15 @@ html_sidebars = {
 }
 html_short_title = "%s-%s" % (project, version)
 
+autoclass_content = "both"
 autodoc_member_order = "bysource"
 
 napoleon_use_ivar = True
 napoleon_use_rtype = False
 napoleon_use_param = False
+
+linkcheck_anchors_ignore = [
+    "^!",
+    # https://github.com/Querela/discord-notifier-bot#bot-creation-etc
+    "^bot-creation-etc$",
+]